+++ examples/plot_decoding_tutorial.py
@@ -84,10 +84,10 @@ print(fmri_masked.shape)
 #
 # The behavioral labels are stored in a CSV file, separated by spaces.
 #
-# We use numpy to load them in an array.
-import numpy as np
+# We use pandas to load them in an array.
+import pandas as pd
 # Load behavioral information
-behavioral = np.recfromcsv(haxby_dataset.session_target[0], delimiter=" ")
+behavioral = pd.read_csv(haxby_dataset.session_target[0], sep=" ")
 print(behavioral)
 
 ###########################################################################
@@ -105,7 +105,7 @@ print(conditions)
 #
 # To keep only data corresponding to faces or cats, we create a
 # mask of the samples belonging to the condition.
-condition_mask = np.logical_or(conditions == b'face', conditions == b'cat')
+condition_mask = conditions.isin(['face', 'cat'])
 
 # We apply this mask in the sampe direction to restrict the
 # classification to the face vs cat discrimination
@@ -184,9 +184,11 @@ from sklearn.cross_validation import KFold
 cv = KFold(n=len(fmri_masked), n_folds=5)
 
 for train, test in cv:
-    svc.fit(fmri_masked[train], conditions[train])
+    conditions_masked = conditions.values[train]
+    svc.fit(fmri_masked[train], conditions_masked)
     prediction = svc.predict(fmri_masked[test])
-    print((prediction == conditions[test]).sum() / float(len(conditions[test])))
+    print((prediction == conditions.values[test]).sum()
+           / float(len(conditions.values[test])))
 
 ###########################################################################
 # Cross-validation with scikit-learn
@@ -278,4 +280,3 @@ show()
 # * :ref:`space_net`
 #
 # ______________
-