+++ doc/whats_new.rst
@@ -24,6 +24,8 @@ Changelog
       :func:`nilearn.plotting.plot_prob_atlas` is deprecated. Use argument
       `bg_img` instead.
 
+- The examples now use pandas for the behavioral information.
+
 Bug fix
 -------